requirements.txt
matplotlib
numba
pandas
plotly.express
pyadi-iio
pylibiio
scipy
pynput